Refactor sky-bench into modules
Also, error handling was improved.
